+++ src/components/Form/VLCTextarea.vue
@@ -1,0 +1,17 @@
+<template>
+    <textarea v-bind="getAttributes"
+              class="vlc-textarea"
+              @input="inputEvent"
+    />
+</template>
+
+<script>
+  import FormControls from "./FormControls";
+  
+  export default {
+    name: "ATBLTextarea",
+    mixins: [
+      FormControls
+    ]
+  }
+</script>