+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.0.0] - 2021-02-14
 ### Added
+- Textarea, Input and Numbers components
 - Switch component
 - Custom font icons
 - Alert component added